% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_EXA.R
\docType{data}
\name{EXA}
\alias{EXA}
\alias{EXA_C01_S04_01}
\alias{EXA_C02_S06_02}
\alias{EXA_C06_S02_04}
\alias{EXA_C07_S02_03}
\alias{EXA_C07_S03_02}
\alias{EXA_C07_S04_01}
\alias{EXA_C07_S07_01}
\alias{EXA_C08_S02_01}
\alias{EXA_C08_S03_01}
\alias{EXA_C08_S04_01}
\alias{EXA_C08_S04_02}
\alias{EXA_C08_S05_02}
\alias{EXA_C09_S03_01}
\alias{EXA_C09_S07_01}
\alias{EXA_C10_S03_01}
\alias{EXA_C10_S06_01}
\alias{EXA_C11_S01_01}
\alias{EXA_C11_S01_02}
\alias{EXA_C11_S02_01}
\alias{EXA_C11_S02_03}
\alias{EXA_C11_S03_01}
\alias{EXA_C11_S04_02}
\alias{EXA_C11_S05_01}
\alias{EXA_C12_S03_03}
\alias{EXA_C13_S03_02}
\alias{EXA_C13_S05_01}
\alias{EXA_C13_S06_01}
\alias{EXA_C13_S07_01}
\alias{EXA_C13_S08_02}
\alias{EXA_C13_S09_01}
\alias{EXA_C13_S09_02}
\alias{EXA_C13_S10_01}
\alias{EXA_C13_S10_02}
\alias{EXA_C14_S03_01}
\alias{EXA_C14_S05_01}
\title{EXA Data}
\format{
An object of class \code{data.frame} with 189 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 1 columns.

An object of class \code{data.frame} with 35 rows and 1 columns.

An object of class \code{data.frame} with 17 rows and 2 columns.

An object of class \code{data.frame} with 10 rows and 2 columns.

An object of class \code{data.frame} with 12 rows and 2 columns.

An object of class \code{data.frame} with 17 rows and 1 columns.

An object of class \code{data.frame} with 144 rows and 2 columns.

An object of class \code{data.frame} with 15 rows and 3 columns.

An object of class \code{data.frame} with 72 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 6 columns.

An object of class \code{data.frame} with 80 rows and 3 columns.

An object of class \code{data.frame} with 109 rows and 3 columns.

An object of class \code{data.frame} with 155 rows and 2 columns.

An object of class \code{data.frame} with 71 rows and 3 columns.

An object of class \code{data.frame} with 29 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 15 rows and 4 columns.

An object of class \code{data.frame} with 100 rows and 4 columns.

An object of class \code{data.frame} with 36 rows and 3 columns.

An object of class \code{data.frame} with 30 rows and 7 columns.

An object of class \code{data.frame} with 184 rows and 2 columns.

An object of class \code{data.frame} with 45 rows and 4 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 12 rows and 3 columns.

An object of class \code{data.frame} with 16 rows and 2 columns.

An object of class \code{data.frame} with 15 rows and 2 columns.

An object of class \code{data.frame} with 36 rows and 1 columns.

An object of class \code{data.frame} with 10 rows and 5 columns.

An object of class \code{data.frame} with 9 rows and 4 columns.

An object of class \code{data.frame} with 16 rows and 5 columns.

An object of class \code{data.frame} with 20 rows and 3 columns.

An object of class \code{data.frame} with 35 rows and 3 columns.

An object of class \code{data.frame} with 39 rows and 4 columns.

An object of class \code{data.frame} with 40 rows and 5 columns.
}
\usage{
EXA_C01_S04_01

EXA_C02_S06_02

EXA_C06_S02_04

EXA_C07_S02_03

EXA_C07_S03_02

EXA_C07_S04_01

EXA_C07_S07_01

EXA_C08_S02_01

EXA_C08_S03_01

EXA_C08_S04_01

EXA_C08_S04_02

EXA_C08_S05_02

EXA_C09_S03_01

EXA_C09_S07_01

EXA_C10_S03_01

EXA_C10_S06_01

EXA_C11_S01_01

EXA_C11_S01_02

EXA_C11_S02_01

EXA_C11_S02_03

EXA_C11_S03_01

EXA_C11_S04_02

EXA_C11_S05_01

EXA_C12_S03_03

EXA_C13_S03_02

EXA_C13_S05_01

EXA_C13_S06_01

EXA_C13_S07_01

EXA_C13_S08_02

EXA_C13_S09_01

EXA_C13_S09_02

EXA_C13_S10_01

EXA_C13_S10_02

EXA_C14_S03_01

EXA_C14_S05_01
}
\description{
EXA Data
}
\keyword{datasets}
\keyword{internal}
