% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_EXR.R
\docType{data}
\name{EXR}
\alias{EXR}
\alias{EXR_C02_S03_01}
\alias{EXR_C02_S03_02}
\alias{EXR_C02_S03_03}
\alias{EXR_C02_S03_04}
\alias{EXR_C02_S03_05}
\alias{EXR_C02_S03_06}
\alias{EXR_C02_S03_07}
\alias{EXR_C02_S03_08}
\alias{EXR_C02_S03_09}
\alias{EXR_C02_S03_11}
\alias{EXR_C02_S03_12}
\alias{EXR_C02_S05_03}
\alias{EXR_C02_S05_06}
\alias{EXR_C06_S02_03}
\alias{EXR_C06_S04_10}
\alias{EXR_C06_S09_05}
\alias{EXR_C06_S10_07}
\alias{EXR_C07_S02_13}
\alias{EXR_C07_S02_17}
\alias{EXR_C07_S02_18}
\alias{EXR_C07_S03_03}
\alias{EXR_C07_S03_04}
\alias{EXR_C07_S03_06}
\alias{EXR_C07_S03_07}
\alias{EXR_C07_S03_08}
\alias{EXR_C07_S03_12}
\alias{EXR_C07_S04_01}
\alias{EXR_C07_S04_02}
\alias{EXR_C07_S04_03}
\alias{EXR_C07_S04_04}
\alias{EXR_C07_S08_07}
\alias{EXR_C08_S02_01}
\alias{EXR_C08_S02_02}
\alias{EXR_C08_S02_03}
\alias{EXR_C08_S02_04}
\alias{EXR_C08_S02_05}
\alias{EXR_C08_S02_07}
\alias{EXR_C08_S02_08}
\alias{EXR_C08_S03_01}
\alias{EXR_C08_S03_02}
\alias{EXR_C08_S03_03}
\alias{EXR_C08_S03_04}
\alias{EXR_C08_S03_05}
\alias{EXR_C08_S04_01}
\alias{EXR_C08_S04_02}
\alias{EXR_C08_S04_03}
\alias{EXR_C08_S04_06}
\alias{EXR_C08_S05_01}
\alias{EXR_C08_S05_02}
\alias{EXR_C08_S05_03}
\alias{EXR_C08_S05_05}
\alias{EXR_C09_S03_02}
\alias{EXR_C09_S03_03}
\alias{EXR_C09_S03_04}
\alias{EXR_C09_S03_06}
\alias{EXR_C09_S03_07}
\alias{EXR_C09_S07_01}
\alias{EXR_C09_S07_02}
\alias{EXR_C09_S07_03}
\alias{EXR_C09_S07_05}
\alias{EXR_C09_S07_06}
\alias{EXR_C10_S03_01}
\alias{EXR_C10_S03_02}
\alias{EXR_C10_S03_03}
\alias{EXR_C10_S03_04}
\alias{EXR_C10_S03_05}
\alias{EXR_C10_S03_06}
\alias{EXR_C10_S06_01}
\alias{EXR_C10_S06_02}
\alias{EXR_C10_S06_03}
\alias{EXR_C10_S06_04}
\alias{EXR_C11_S02_01}
\alias{EXR_C11_S02_02}
\alias{EXR_C11_S02_03}
\alias{EXR_C11_S02_04}
\alias{EXR_C11_S03_01}
\alias{EXR_C11_S03_02}
\alias{EXR_C11_S03_03}
\alias{EXR_C11_S04_01}
\alias{EXR_C11_S04_02}
\alias{EXR_C11_S05_01}
\alias{EXR_C11_S05_02}
\alias{EXR_C13_S03_03}
\alias{EXR_C13_S05_01}
\alias{EXR_C13_S05_02}
\alias{EXR_C13_S06_01}
\alias{EXR_C13_S06_02}
\alias{EXR_C13_S06_03}
\alias{EXR_C13_S07_02}
\alias{EXR_C13_S07_03}
\alias{EXR_C13_S08_01}
\alias{EXR_C13_S08_02}
\alias{EXR_C13_S08_03}
\alias{EXR_C13_S08_04}
\alias{EXR_C13_S08_05}
\alias{EXR_C13_S08_06}
\alias{EXR_C13_S09_01}
\alias{EXR_C13_S09_02}
\alias{EXR_C13_S09_03}
\alias{EXR_C13_S10_01}
\alias{EXR_C13_S10_02}
\alias{EXR_C13_S10_03}
\alias{EXR_C13_S10_04}
\alias{EXR_C13_S10_05}
\alias{EXR_C13_S10_06}
\alias{EXR_C14_S03_01}
\alias{EXR_C14_S03_02}
\alias{EXR_C14_S04_02}
\title{EXR Data}
\format{
An object of class \code{data.frame} with 90 rows and 1 columns.

An object of class \code{data.frame} with 159 rows and 1 columns.

An object of class \code{data.frame} with 29 rows and 1 columns.

An object of class \code{data.frame} with 53 rows and 1 columns.

An object of class \code{data.frame} with 45 rows and 1 columns.

An object of class \code{data.frame} with 60 rows and 1 columns.

An object of class \code{data.frame} with 155 rows and 1 columns.

An object of class \code{data.frame} with 30 rows and 1 columns.

An object of class \code{data.frame} with 50 rows and 2 columns.

An object of class \code{data.frame} with 216 rows and 1 columns.

An object of class \code{data.frame} with 109 rows and 1 columns.

An object of class \code{data.frame} with 30 rows and 1 columns.

An object of class \code{data.frame} with 20 rows and 1 columns.

An object of class \code{data.frame} with 16 rows and 1 columns.

An object of class \code{data.frame} with 32 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 1 columns.

An object of class \code{data.frame} with 26 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 1 columns.

An object of class \code{data.frame} with 50 rows and 1 columns.

An object of class \code{data.frame} with 21 rows and 1 columns.

An object of class \code{data.frame} with 63 rows and 2 columns.

An object of class \code{data.frame} with 174 rows and 2 columns.

An object of class \code{data.frame} with 24 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 82 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.

An object of class \code{data.frame} with 15 rows and 2 columns.

An object of class \code{data.frame} with 66 rows and 2 columns.

An object of class \code{data.frame} with 11 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.

An object of class \code{data.frame} with 23 rows and 2 columns.

An object of class \code{data.frame} with 329 rows and 2 columns.

An object of class \code{data.frame} with 96 rows and 2 columns.

An object of class \code{data.frame} with 113 rows and 2 columns.

An object of class \code{data.frame} with 164 rows and 2 columns.

An object of class \code{data.frame} with 29 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 178 rows and 2 columns.

An object of class \code{data.frame} with 96 rows and 3 columns.

An object of class \code{data.frame} with 10 rows and 5 columns.

An object of class \code{data.frame} with 20 rows and 3 columns.

An object of class \code{data.frame} with 16 rows and 3 columns.

An object of class \code{data.frame} with 12 rows and 3 columns.

An object of class \code{data.frame} with 40 rows and 3 columns.

An object of class \code{data.frame} with 35 rows and 3 columns.

An object of class \code{data.frame} with 48 rows and 3 columns.

An object of class \code{data.frame} with 20 rows and 6 columns.

An object of class \code{data.frame} with 24 rows and 3 columns.

An object of class \code{data.frame} with 72 rows and 3 columns.

An object of class \code{data.frame} with 44 rows and 3 columns.

An object of class \code{data.frame} with 13 rows and 3 columns.

An object of class \code{data.frame} with 10 rows and 2 columns.

An object of class \code{data.frame} with 17 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 22 rows and 2 columns.

An object of class \code{data.frame} with 27 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.

An object of class \code{data.frame} with 90 rows and 2 columns.

An object of class \code{data.frame} with 18 rows and 2 columns.

An object of class \code{data.frame} with 30 rows and 2 columns.

An object of class \code{data.frame} with 15 rows and 2 columns.

An object of class \code{data.frame} with 35 rows and 3 columns.

An object of class \code{data.frame} with 100 rows and 4 columns.

An object of class \code{data.frame} with 10 rows and 3 columns.

An object of class \code{data.frame} with 20 rows and 3 columns.

An object of class \code{data.frame} with 25 rows and 3 columns.

An object of class \code{data.frame} with 20 rows and 8 columns.

An object of class \code{data.frame} with 40 rows and 4 columns.

An object of class \code{data.frame} with 12 rows and 3 columns.

An object of class \code{data.frame} with 15 rows and 3 columns.

An object of class \code{data.frame} with 15 rows and 6 columns.

An object of class \code{data.frame} with 44 rows and 3 columns.

An object of class \code{data.frame} with 100 rows and 3 columns.

An object of class \code{data.frame} with 17 rows and 3 columns.

An object of class \code{data.frame} with 90 rows and 3 columns.

An object of class \code{data.frame} with 100 rows and 8 columns.

An object of class \code{data.frame} with 35 rows and 7 columns.

An object of class \code{data.frame} with 68 rows and 7 columns.

An object of class \code{data.frame} with 4 rows and 3 columns.

An object of class \code{data.frame} with 184 rows and 2 columns.

An object of class \code{data.frame} with 45 rows and 4 columns.

An object of class \code{data.frame} with 45 rows and 5 columns.

An object of class \code{data.frame} with 15 rows and 3 columns.

An object of class \code{data.frame} with 30 rows and 2 columns.

An object of class \code{data.frame} with 30 rows and 2 columns.

An object of class \code{data.frame} with 70 rows and 2 columns.

An object of class \code{data.frame} with 17 rows and 2 columns.

An object of class \code{data.frame} with 83 rows and 2 columns.

An object of class \code{data.frame} with 30 rows and 1 columns.

An object of class \code{data.frame} with 20 rows and 1 columns.

An object of class \code{data.frame} with 232 rows and 2 columns.

An object of class \code{data.frame} with 15 rows and 2 columns.

An object of class \code{data.frame} with 53 rows and 2 columns.

An object of class \code{data.frame} with 22 rows and 2 columns.

An object of class \code{data.frame} with 44 rows and 2 columns.

An object of class \code{data.frame} with 22 rows and 3 columns.

An object of class \code{data.frame} with 9 rows and 4 columns.

An object of class \code{data.frame} with 15 rows and 11 columns.

An object of class \code{data.frame} with 10 rows and 6 columns.

An object of class \code{data.frame} with 15 rows and 3 columns.

An object of class \code{data.frame} with 10 rows and 3 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.

An object of class \code{data.frame} with 20 rows and 2 columns.

An object of class \code{data.frame} with 30 rows and 2 columns.

An object of class \code{data.frame} with 17 rows and 3 columns.

An object of class \code{data.frame} with 53 rows and 3 columns.

An object of class \code{data.frame} with 62 rows and 2 columns.

An object of class \code{data.frame} with 50 rows and 4 columns.
}
\usage{
EXR_C02_S03_01

EXR_C02_S03_02

EXR_C02_S03_03

EXR_C02_S03_04

EXR_C02_S03_05

EXR_C02_S03_06

EXR_C02_S03_07

EXR_C02_S03_08

EXR_C02_S03_09

EXR_C02_S03_11

EXR_C02_S03_12

EXR_C02_S05_03

EXR_C02_S05_06

EXR_C06_S02_03

EXR_C06_S04_10

EXR_C06_S09_05

EXR_C06_S10_07

EXR_C07_S02_13

EXR_C07_S02_17

EXR_C07_S02_18

EXR_C07_S03_03

EXR_C07_S03_04

EXR_C07_S03_06

EXR_C07_S03_07

EXR_C07_S03_08

EXR_C07_S03_12

EXR_C07_S04_01

EXR_C07_S04_02

EXR_C07_S04_03

EXR_C07_S04_04

EXR_C07_S08_07

EXR_C08_S02_01

EXR_C08_S02_02

EXR_C08_S02_03

EXR_C08_S02_04

EXR_C08_S02_05

EXR_C08_S02_07

EXR_C08_S02_08

EXR_C08_S03_01

EXR_C08_S03_02

EXR_C08_S03_03

EXR_C08_S03_04

EXR_C08_S03_05

EXR_C08_S04_01

EXR_C08_S04_02

EXR_C08_S04_03

EXR_C08_S04_06

EXR_C08_S05_01

EXR_C08_S05_02

EXR_C08_S05_03

EXR_C08_S05_05

EXR_C09_S03_02

EXR_C09_S03_03

EXR_C09_S03_04

EXR_C09_S03_06

EXR_C09_S03_07

EXR_C09_S07_01

EXR_C09_S07_02

EXR_C09_S07_03

EXR_C09_S07_05

EXR_C09_S07_06

EXR_C10_S03_01

EXR_C10_S03_02

EXR_C10_S03_03

EXR_C10_S03_04

EXR_C10_S03_05

EXR_C10_S03_06

EXR_C10_S06_01

EXR_C10_S06_02

EXR_C10_S06_03

EXR_C10_S06_04

EXR_C11_S02_01

EXR_C11_S02_02

EXR_C11_S02_03

EXR_C11_S02_04

EXR_C11_S03_01

EXR_C11_S03_02

EXR_C11_S03_03

EXR_C11_S04_01

EXR_C11_S04_02

EXR_C11_S05_01

EXR_C11_S05_02

EXR_C13_S03_03

EXR_C13_S05_01

EXR_C13_S05_02

EXR_C13_S06_01

EXR_C13_S06_02

EXR_C13_S06_03

EXR_C13_S07_02

EXR_C13_S07_03

EXR_C13_S08_01

EXR_C13_S08_02

EXR_C13_S08_03

EXR_C13_S08_04

EXR_C13_S08_05

EXR_C13_S08_06

EXR_C13_S09_01

EXR_C13_S09_02

EXR_C13_S09_03

EXR_C13_S10_01

EXR_C13_S10_02

EXR_C13_S10_03

EXR_C13_S10_04

EXR_C13_S10_05

EXR_C13_S10_06

EXR_C14_S03_01

EXR_C14_S03_02

EXR_C14_S04_02
}
\description{
EXR Data
}
\keyword{datasets}
\keyword{internal}
